Move TopologyEntry into its own header file
It is needed in the kernel and this consolidates versions of the
structure.  This change also wraps types in types.h that cannot be
used in the kernel in an #ifdef that removes them from compilation
during the kext build.
