+++ src/MacMSRDriver/PcmMsr/UserKernelShared.h
@@ -1,12 +1,20 @@
 // SPDX-License-Identifier: BSD-3-Clause
 // Copyright (c) 2012, Intel Corporation
 // written by Austen Ott
-//    
+//
 #define PcmMsrDriverClassName com_intel_driver_PcmMsr
 #define kPcmMsrDriverClassName "com_intel_driver_PcmMsr"
+
 #ifndef USER_KERNEL_SHARED
 #define USER_KERNEL_SHARED
+
+#define PCM_API
+
 #include <stdint.h>
+#include "../../topologyentry.h"
+
+using namespace pcm;
+
 typedef struct {
     uint64_t value;
     uint32_t cpu_num;
@@ -19,19 +27,6 @@ typedef struct {
     bool mask;
     char padding[115];
 } k_pcm_msr_data_t;
-
-// The topologyEntry struct that is used by PCM
-typedef struct
-{
-    int32_t os_id;
-    int32_t thread_id;
-    int32_t core_id;
-    int32_t tile_id;
-    int32_t socket;
-    int32_t native_cpu_model;
-    int32_t core_type; // This is an enum in the userland structure.
-    int32_t padding;
-} topologyEntry;
 
 enum {
     kOpenDriver,
