+++ src/cpucounters.h
@@ -23,6 +23,7 @@
 #undef PCM_UNCORE_PMON_BOX_CHECK_STATUS // debug only
 
 #include "types.h"
+#include "topologyentry.h"
 #include "msr.h"
 #include "pci.h"
 #include "bw.h"
@@ -85,38 +86,6 @@ class SystemRoot;
 
         A set of performance monitoring routines for recent Intel CPUs
 */
-
-struct PCM_API TopologyEntry // describes a core
-{
-    int32 os_id;
-    int32 thread_id;
-    int32 core_id;
-    int32 tile_id; // tile is a constalation of 1 or more cores sharing salem L2 cache. Unique for entire system
-    int32 socket;
-    int32 native_cpu_model = -1;
-    enum CoreType
-    {
-        Atom = 0x20,
-        Core = 0x40,
-        Invalid = -1
-    };
-    CoreType core_type = Invalid;
-
-    TopologyEntry() : os_id(-1), thread_id (-1), core_id(-1), tile_id(-1), socket(-1) { }
-    const char* getCoreTypeStr()
-    {
-        switch (core_type)
-        {
-            case Atom:
-                return "Atom";
-            case Core:
-                return "Core";
-            case Invalid:
-                return "invalid";
-        }
-        return "unknown";
-    }
-};
 
 class HWRegister
 {
